requirements.txt
asgiref==3.5.0
black==22.1.0
click==8.0.4
crispy-bootstrap5==0.6
Django==4.0.2
django-autoslug==1.9.8
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-markdownx==4.0.0b1
django-taggit==2.1.0
Markdown==3.3.6
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
psycopg2-binary==2.9.3
Pygments==2.11.2
sqlparse==0.4.2
tomli==2.0.1
